requirements.txt
Flask==3.1.2
pandas==2.3.3
numpy>=1.26.0
scipy>=1.12.0
scikit-learn==1.8.0
supabase>=2.27.0
python-dotenv==1.0.0
websockets>=11,<16
gunicorn==21.2.0

